Added CurrencyHelper, moved currency values to use ints

+++ LotteryGame.Shared/GameResult.cs
@@ -1,7 +1,7 @@
 ﻿namespace LotteryGame.Shared {
     public class GameResult {
         public Dictionary<string, List<int>> Winners = new Dictionary<string, List<int>>();
-        public Dictionary<string, double> Prizes = new Dictionary<string, double>();
-        public double HouseShare { get; set; }
+        public Dictionary<string, int> Prizes = new Dictionary<string, int>();
+        public int HouseShare { get; set; }
     }
 }
